moved TodoItem components directory
